Cleaning up PeerConnectionFactory dependency injection.
This CL uses a "PeerConnectionFactoryDependencies" struct to
encapsulate all possible injectable dependencies, optional or
mandatory. Instead of adding new CreatePeerConnectionFactory overloads
when we want something new to become injectible, we can just add a new
member to this struct.

The API user can call "PeerConnectionFactoryDependencies::Create()" to
get a struct with default implementations provided for all mandatory
dependencies (currently none), and then override them with their own if
desired.

Pulling out new injectable modules that were previously always built-in
will be slightly cleaner with this approach; we can:
1. Add a new member to the struct.
2. Make the existing "Create" method create the default implementation,
   so existing applications keep working as before.
3. Add a new "Create" method that will *not* create the default
   implementation (would be in a separate build target).
4. Eventually, deprecate the old Create method and force applications
   to depend on the new module explicitly.

This isn't fundamentally different to adding new
CreatePeerConnectionFactory overloads, but the code ends up being
cleaner.

This CL also allows passing in null audio encoder/decoder factories,
since it would be odd if
"CreatePeerConnectionFactory(PeerConnectionFactoryDependencies::Create())"
failed, and it's inconsistent that passing in null factories is allowed
for video but not audio.

Bug: webrtc:7913
Change-Id: Ic16ca5dc81899c7d7a8639e16086df87eeeffd81