Fix a bug in TurnServer that causes flakiness in webrtc_perf_tests.
When a TCP TURN port is destroyed, a TURN refresh request with zero
lifetime is first sent to release the TURN allocation at the server, and
the underlying TCP connection is closed afterwards.

The closing of the TCP connection is handled first by the
VirtualSocketServer in our test infrastructure, and the corresponding
server socket is asynchronously destroyed at the TURN server. The
refresh request is however still passed to this server socket, which
fails a DCHECK. The implementation of the TURN server should disable any
firing of signals from a server socket to destroy.

Putting None to the bug id since this is one-liner CL.

Bug: None
Change-Id: Ib457b3800511a322ef69d67c71f2de05f3d67967