Update patch set 1
Patch Set 1:

Hello all. This is my first CL and I've tried my best to navigate the process and not waste anyone's time. Tips for improvement are very welcome.

This CL adds two bits of functionality which were still missing in iOS. The ability to generate a new RTCCertificate, and the ability to (re)use a specific RTCCertificate when creating an RTCPeerConnection. JS already supports this but iOS does not and instead automatically generates an RTCCertificate for every connection.

Thanks again for your time and work on this project.

Patch-set: 1
Reviewer: Gerrit User 7443 <7443@58829da1-049c-39fb-b951-ebdcd0984767>
Work-in-progress: false