Remove not-updating-stats log message.
GetStats() is currently the only way to get smoothed audio levels. It's
expected to be called quite frequently. We use 100ms intervals, so a
call to GetStats() for any reason other than audio levels has a 50%
chance of triggering this log line. This makes it too noisy for LS_INFO.

The log line was added recently
(https://webrtc-review.googlesource.com/c/src/+/82260) and doesn't seem
very useful for diagnostic purposes, so remove it entirely.

Bug: n/a
Change-Id: I15700085c60b9929a4df2e2327012a4f16b505b6