Update patch set 1
Patch Set 1:

(3 comments)

Hi Michael,

Thanks for the patch! I think this is a good start and will hopefully help other app developers who wish to specify a particular certificate.

For me the most important change is to break out RTCCertificate into a proper wrapper (like RTCRtpParameters, for example). That will allow us to extend it more in the future and prevent the awkward generate -> export to PEM -> reparse from PEM -> feed it to config which is what would happen now.

+andersc for iOS expertise/owners.

Patch-set: 1
CC: Gerrit User 5039 <5039@58829da1-049c-39fb-b951-ebdcd0984767>