Update patch set 18
Patch Set 18:

(22 comments)

Here's a bunch of comments! For the IAudioClient versioning, we can either explicitly keep an object of the correct type (i.e. IAudioClient3) and have the right version for the right functionality in our interfaces (see a couple of comments on core_audio_utility_win.h), or we could just always hold an IAudioClient and use the Windows version to determine which version is really implemented. Currently, there's a bit of both.

I've not thought heavily about the restart logic, but I'm not certain it's not racy. Please explain a bit more. I'll look over it again in a bit.

Patch-set: 18
Reviewer: Gerrit User 5399 <5399@58829da1-049c-39fb-b951-ebdcd0984767>