Update patch set 5
Patch Set 5:

> Patch Set 4:
> 
> The scaled image after the change does not seem to be entirely correct either. There is a border of 1-2 pixels on the right and bottom edges that is off.
> 
> Can you explain why this happens?

I think the reason for the border on the right and bottom edge is because when I created that image I created a RTCCVPixelBuffer with a cropX of 2 and cropY of 3 but I passed in the original height and width as the cropWidth and cropHeight. As a result, the borders you mentioned are due to "zeroed out" pixels due to the crop but are still visually visible because the cropWidth and cropHeight are the same as the original image. I changed the test to correctly pass in the appropriate cropWidth and cropHeight.

I tried this again with the correct cropWidth and cropHeight with a cropX of 2 and cropY of 3 and this was the result (there is no border anymore): [https://i.imgur.com/4euFOyX.png]

To see it a bit more easily here is what it looks like with a cropX of 200 and cropY of 300 with the cropWidth and cropHeight set correctly: [https://i.imgur.com/Z7JyX7G.png]. It looks like the top right red gradient and bottom left white gradient are cropped out correctly due to the large cropX and cropY values.

Patch-set: 5
Reviewer: Gerrit User 7527 <7527@58829da1-049c-39fb-b951-ebdcd0984767>