[Window capture] filter out sibling windows with same title.
On Windows, the Window Media Player window seems consisting of several
sibling windows, with same window title as the main window and from same
process. Currently CroppingWindowCapturerWin will think the WMPlayer
window is overlapped by those windows and switch to GDI capture method,
which is not supported by WMPlayer. So Chrome can't capture and share
the content played in WMPlayer.

This cl is to extend the current null title check to include the case
that window has same title as the selected window.

Bug: chromium:865193
Change-Id: Id16b21596ab3b870197758679e5406138ac1a432