Update patch set 6
Patch Set 6:

(4 comments)

I am sorry that it took me a few days of detour before coming back to this CL.

I have not made code changes, mostly because of the discussion below. The two open comments from the last iteration of this CL is:
1. We have unregistered attribute type values in our ICE STUN messages, and STUN_ATTR_NETWORK_INFO is already conflicting with one registered with IANA. Now we are introducing one more unregistered attribute STUN_ATTR_NETWORK_INFO_EXTRA.
2. Related to one, STUN_ATTR_NETWORK_INFO_EXTRA can be combined with STUN_ATTR_NETWORK_INFO to a) save 4-byte overhead b) remove one unregistered attribute

For 1, my thought is that the usage of these non-standard attributes in ICE STUN is limited to our implementation, and that is probably the reason our codebase has introduced these existing non-standard attributes with on registration, though I agree this is not ideal.

For 2, it is indeed reasonable to me to do it, but it seems -- correct me if I am wrong -- for interop with old endpoints that only understand a shorter bit-width (32 bits) of STUN_ATTR_NETWORK_INFO, we need to tweak the byte layout in the new type, say a 64-bit type as in my reply.

I am debating with myself which way to go and please let me know what you think.

Patch-set: 6
Reviewer: Gerrit User 5887 <5887@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0