Changes in response to reviewer comments
Changes in response to reviewer comments


Changes in response to reviewer comments


Changes in response to reviewer comments


AEC3: Increasing the accuracy of the detection for early reverb

This CL introduces an adaptive estimation of the early reverb
in the estimation for the room reverberation. The benefits of
this is that for room with long early reflections there is
less risk of underestimating the reverberation.

Change-Id: Ib88a92856e2c130452744728bc5bdd13e4f3006a
Bug: webrtc:9479, chromium:865397