Update patch set 2
Patch Set 2:

> Patch Set 2: Commit-Queue+1
> 
> PTAL. Do you think this makes sense? As far as I understand, 90 kHz is kind-of the default frequency in RTP, and different values are used only by some audio codecs (to match sample frequency, I guess).
> 
> Having a default value simplifies handling of video packets, since we don't need to worry about calling set_payload_type_frequency at the right place.
> 
> Later on, I would like payload_type_frequency to be set where the packets are parsed, using some per-rtp-session mapping of payload type to frequency, but that's not yet an easy change.

I do not feel this value belong to the RtpPacket[Received]

90kHz is default _video_ rtp frequency.

The proper solution is to use whatever frequency is configured for given payload type in the sdp (that field is present for all payload types)

Until then the old solution: set it to k_Video_PaylodTypeFrequency in Rtp_Video_StreamReceiver, i.e. assume it is always 90k for video packets makes more sense to me.
Many places of video pipeline assumes it is the only possible value.

Alternative reasonable solution is to ignore the rtp frequency in the video pipeline.

Patch-set: 2
Reviewer: Gerrit User 5019 <5019@58829da1-049c-39fb-b951-ebdcd0984767>