Update patch set 2
Patch Set 2:

> The proper solution is to use whatever frequency is configured for given payload type in the sdp (that field is present for all payload types)

Right.

> Until then the old solution: set it to k_Video_PaylodTypeFrequency in Rtp_Video_StreamReceiver, i.e. assume it is always 90k for video packets makes more sense to me.

Problem is, RtpVideoStreamReceiver::OnRtpPacket get a const RtpPacketReceived&. So we'd need to make a copy to modify the field there. (We'll have the same problem on the audio side, but I'm not touching that code at the moment).

An alternative for the time being might be to set it where the RtpPacketReceived is currently created, in Call::DeliverRtp and Call:RecoveredPacket.
 
> Alternative reasonable solution is to ignore the rtp frequency in the video pipeline.

I'm aware of one use, where the packet is passed to StreamStatistician::IsRetransmitOfOldPacket. That interface currently uses an RTPHeader rather than an RtpPacketReceived, so for now it's no problem to set the frequency there.

Patch-set: 2
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0