Update patch set 2
Patch Set 2:

The implementation looks perfect. But, I have a question about the timing of the timestamp. It seems to me that the timestamp is written when APM is initialized. Is it possible that there is an arbitrary delay from initialization until media starts flowing (and is recorded)? Would this delay be known?

Patch-set: 2
Reviewer: Gerrit User 5100 <5100@58829da1-049c-39fb-b951-ebdcd0984767>