Declare targets in desktop_capture/BUILD.gn only on supported platforms
This file has targets that work only on desktop platforms, but they are not wrapped in a conditional.
This never caused any problems because GN just ignores files if nothing happens to depend on targets that are in them.
But it's still good to be explicit about it.

This un-nests conditionals like `if(desktop_capture_generic){sources += ...}`
and you would think that this file used to be nice by exposing non-platform-specific files to unittests that can run on any platform.
But none of that matters because the file was just ignored on those platforms.

Bug: None
Change-Id: Iea85a2ce12eba4bd64a6fed23412846997a648d4