Simulcast screenshare adjustment to temporal layers, bitrate
Change experimental max bitrate setting from 1.6Mbps to 1.25Mbps in
order to allow a larger fraction of participants to receive this layer.

Add a new field trial to allow setting the number of temporal layers for
the high-quality simulcast stream in screensharing separately from the
temporal layer count for regular video.

Bug: webrtc:9477
Change-Id: I1341b774f870c50710901da24963bd3ede96ffd8