Update patch set 2
Patch Set 2:

If possible, it would be preferable to move the check higher up the call chain, and not invoke these methods at all before we have a known non-zero frame size.

Where are the zero size coming from? Are there calls to UpdateFrameSize with zero inputs, or are the zeros from construction?

Patch-set: 2
CC: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>