Adds support for JSON config in video_replay.
This change adds the ability to optionally provide configuration to the
video_replay binary through JSON. This is a merge of the changes provided by
philipel on this bug: https://bugs.chromium.org/p/chromium/issues/detail?id=840536#c1

However it has been updated to pull all the json parsing into the example binary
itself instead of it being integrated into the core library. Writing test cases
out to JSON configuration will be handled in a different CL. Most likely there
will be a utility class added that takes a Config and converts it to JSON that is
decoupled from the actual implementation.

Bug: webrtc:9609
Change-Id: Icc5900063d7f704825f224240e4b3787c06ca074