Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Hi,
> 
> This is the chain you should follow:
> 
> FixedGainController::SetGain() -> GainCurveApplier::Reset() -> FixedDigitalLevelEstimator::Reset() -> FixedDigitalLevelEstimator::filter_state_level_ = 0.f
> 
> I did test this change in a downstream project.
> Could you check again?
> 
> Alessio

Sorry, I didn't look closely! I thought 'filter_state_level_' was a new variable that was never read and only written to. Now I see it works.

LG, but I want a test on this. The behavior that we are looking for is easy to test.

Patch-set: 2
Reviewer: Gerrit User 5124 <5124@58829da1-049c-39fb-b951-ebdcd0984767>