AEC3: Simplify render buffering
This CL simplifies the buffering of render data. It removes the concepts
of call skew and maximum jitter headroom. Instead of making assumptions
about the worst possible platform, it leverages recent improvements in
the delay estimator to quickly adapt when the conditions change.

Pros:
- Cleaner code that makes the concept of delay more clear.
- No capture delay, delay is found ~200 ms faster.
- Allows for removal of one matched filter because of the jitter headroom
removal.

Cons:
- Delay estimator needs to re-adapt when the call jitter increases.

The code can be deactivated by a kill switch. When the kill switch is
pulled the CL is bit exact.

Change-Id: Ie2f9c8c5ce5b5a4510b4bdb95db2b970b57cd5d0