Update patch set 3
Patch Set 3: Commit-Queue+2

> Patch Set 3: Code-Review+1
> 
> Nothing wrong with this CL in particular, but I'm getting the sense that we're repeating the same code for arithmetic and infinities in a number of places...

Yes I agree, also I think Timestamp and TimeDelta seems to get more similar to each other. At the same time, the fact that they are different and have defined and strict semantics for conversion between each other makes it useful to not use the same underlying class.

It might be possible to do something with templates and inheritance, but the risk is that it becomes really hard to understand. Just having a shared common implementation and wrap the interface could work, but I'm afraid that the wrapper would be about as verbose and mostly result in more indirection.

Patch-set: 3
Reviewer: Gerrit User 5531 <5531@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=+2