Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1

Thanks for the review. This is my first contribution here and I am kinda lost. I am not a commiter so I could not run tryjobs. When is the step of running building and testing the CL supposed to happen?. The doc (https://webrtc.org/contributing/) does not mention anything about it. It just says _"Once the reviewer LGTMs (approves) the patch, ask them to put it into the commit queue."_ so should I ask that explicitly? Thanks.

Patch-set: 1
Reviewer: Gerrit User 7742 <7742@58829da1-049c-39fb-b951-ebdcd0984767>