Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> This wasn't easy to catch. I am wondering if it makes sense to have a test to ensure the framework is genuine after the generation.

Yeah I agree. The macOS version of AppRTCMobile does link with the framework, and it is built by the bots. But I think this was also a question of the current state of the macos build folder when building the framework, if there was already a Headers/ symlink in the right location generated by an earlier build, then it was working I believe.

Patch-set: 1
Reviewer: Gerrit User 5249 <5249@58829da1-049c-39fb-b951-ebdcd0984767>