Update patch set 1
Patch Set 1:

(1 comment)

I'm wondering, couldn't we just as well just not enable enforce_min_bitrate for strams with alr_probing on? Or is it used in more places?

I think this could be in the right direction. With this we could also move the final decision to use ALR probing away from each media stram, which currently is the case. I wonder if we shouldn't disable enforce for all strams if alr probing is enabled though. With this change we could still have alr probing enabled and padding enabled.

I'd prefer that this kinds of decisions were moved away from each media stream and into the bitrate estimator. But I got push-back on that idea in the past so this might

Patch-set: 1
Reviewer: Gerrit User 5531 <5531@58829da1-049c-39fb-b951-ebdcd0984767>