Update patch set 1
Patch Set 1:

This cl is far from ready, but maybe you have some advice on an early stage. Do you see any natural way to split this up in pieces, or some cleanups that it would help to do first?

To get it in shape, my next steps are to delete no-longer used code from voe::Channel, and figure out what locks we need to move along.

One problem is that the receive pipeline seems to be

rtp-specific code ---> AudioCodingModule ---> more rtp-specifc code.

For now, the first part is moved to RtpAudioStreamReceiver, and the latter part (mainly timestamps) is moved to AudioReceiveStream, where it doesn't belong. The intended direction is to let RtpAudioStreamreceiver be unaware of audio decoding and other audio processing, and let AudioReceiveStream to be unaware of RTP.

Patch-set: 1
Reviewer: Gerrit User 5077 <5077@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>
Work-in-progress: false