Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Looks like Clangs return value optimization can handle these cases without help: https://godbolt.org/z/18LLoa . Looks like the Release variant might end up with a few extra instructions to create the empty string.

If enough stuff is inlined so that it can see that it's constructing a new string on the stack, filling it with stuff, returning a copy, and then deleting the original---then yes.

You could try this variant of Release:

    std::string Release2() && {
        std::string ret = std::move(s_);
        s_.clear();
        return ret;
    }

It uses the move constructor instead of swap, and it appears that clang is happy to inline the former but not the latter.

Patch-set: 3
Reviewer: Gerrit User 5126 <5126@58829da1-049c-39fb-b951-ebdcd0984767>