Update patch set 2
Patch Set 2: Code-Review+1

LGTM
I'm usually for the fail-fast rule. But there, that is for logging purpose, and it might be more convenient to see an "Inf" or "Nan" in the json (*) that to let the test crash altogether.

Also, I thing the issue at hand here it the absence of more explicit error message in the subsequent steps:
  * Why did the parsing failed (which line+column).
  * Error message (and stack trace) properly lifted for the world to see.

(*) We might use string in order for the json to be valid.

Patch-set: 2
Reviewer: Gerrit User 7285 <7285@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Code-Review=+1