Remove echo_cancellation() and echo_control_mobile() interface access outside APM
Some of the AEC settings in WebRtcVoiceEngine agree with just about everywhere
else and have therefore been set in stone inside AEC2/AECM. A lot of routing
for those settings disappears.

The comfort noise setting is exposed in the API, so the flag for it will be
removed a PSA later.

Bug: webrtc:9535
Change-Id: I53816152415a9a069cea9520cec697b6bcfe0948