Update patch set 1
Patch Set 1:

Looks reasonable to me, but I wonder if it might be better to move this experiment logic to the rtp send stream configuration in webrtcvideoengine (I think WebRtcVideoChannel::GetChangedSendParameters might be the right place), and arrange so that ulpfec_payload_type is always -1 for this experiment.

Then we can more confident that we get consistent behavior everywhere ulpfec_payload_type is used, and we get the experiment logic closer to related flexfec experiments, WebRTC-FlexFEC-03 and WebRTC-FlexFEC-03-Advertised.

Patch-set: 1
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>