Update patch set 9
Patch Set 9:

(8 comments)

> Patch Set 1:
> 
> This cl is far from ready, but maybe you have some advice on an early stage. Do you see any natural way to split this up in pieces, or some cleanups that it would help to do first?
> 
> To get it in shape, my next steps are to delete no-longer used code from voe::Channel, and figure out what locks we need to move along.
> 
> One problem is that the receive pipeline seems to be
> 
> rtp-specific code ---> AudioCodingModule ---> more rtp-specifc code.
> 
> For now, the first part is moved to RtpAudioStreamReceiver, and the latter part (mainly timestamps) is moved to AudioReceiveStream, where it doesn't belong. The intended direction is to let RtpAudioStreamreceiver be unaware of audio decoding and other audio processing, and let AudioReceiveStream to be unaware of RTP.


Could you make this part of the CL description? it's useful to have this context in the future, and if this is just a comment in the CL, it will be impossible to find it.

I also left some comments that would help me understand better the code; also - we need to decide how we want to get the Rtt & BWE from MediaTransportInterface.

Patch-set: 9