Update patch set 17
Patch Set 17:

> Patch Set 17:
> 
> Big change. Would it be possible to do in steps - e.g. create the RtpReceiver in AudioReceiveStream and push it to voe::Channel as a first step, then progressively move functionality into AudioReceiveStream?

Hmm. I can give that a try. Do you have any suggestion for what to move first?

I have another high-level question: What kind of tests do we need for AudioReceiveStream, after this change? Currently, it delegates a lot of work to ChannelProxy, and then there a tests using a mock ChannelProxy to verify that it makes the expected calls. One could try to mirror that with a mock RtpAudioStreamReceiver, but I'm not sure it's that useful.

Patch-set: 17
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>