Fix a bug that BasicPortAllocator is created on the signaling thread.
PortAllocator allows its construction and configuration on a
non-network thread before Initialize is called. However,
its subclass BasicPortAllocator creates and starts the network monitor
when it is constructed, and the lifecyle of the network monitor should
be restricted to a single thread. As a result, BasicPortAllocator should
only be constructed on the network thread.

Bug: webrtc:9112
Change-Id: I79c31f1a7cd299dad8f9034cc9b83fcd8d3328f7